working on seed files, finished a big chunk

+++ migrations/20161005143028_grocery_list_item.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('grocery_list_item', function(table){
     table.increments();
     table.integer('gl_id').references('grocery_list.id')
-    table.integer('amount');
+    table.float('amount');
     table.text('measurement');
     table.string('name');
   })
